Split board and controls
