+++ src/components/board/Board.jsx
@@ -1,63 +1,10 @@
 import PropTypes from 'prop-types';
 import Cell from "../cell/Cell";
-import { useRef, useState } from 'react';
-import { generateMatrix } from '../../utils/matrix.util';
-import { getPopulation, nextGeneration } from '../../utils/game-of-life.js';
+import { getPopulation } from '../../utils/game-of-life.js';
 import Step from "../../assets/icons/MdiSkipNext.svg";
-import Play from "../../assets/icons/MdiPlay.svg";
-import Pause from "../../assets/icons/MdiPause.svg";
 import styles from "./board.module.scss";
 
-const getSpeed = (sliderValue) => {
-  const speedMs = -20 * sliderValue + 2000;
-  return speedMs;
-}
-
-const Board = () => {
-
-  const [boardState, setBoardState] = useState(generateMatrix(80, 130));
-  const [generationState, setGenerationState] = useState(0);
-  const [playIconState, setPlayIconState] = useState(Play);
-  const [speedState, setSpeedState] = useState(100);
-
-  const intervalRef = useRef();
-
-  const handleCellClick = (rowPosition, columnPosition) => {
-    const update = [...boardState];
-    update[rowPosition][columnPosition] = !update[rowPosition][columnPosition]
-    setBoardState(update);
-  }
-
-  const handlePlay = () => {
-    if( playIconState === Play ) {
-      setPlayIconState(Pause);
-      const speedMs = getSpeed(speedState);
-      intervalRef.current = setInterval( ()=> setNextGeneration(), speedMs);
-    } else {
-      setPlayIconState(Play);
-      clearInterval(intervalRef.current);
-    }
-  }
-
-  const handleNextGeneration = () => setNextGeneration();
-
-  const setNextGeneration = () => {
-    setBoardState(board => nextGeneration(board));    
-    setGenerationState(current => current + 1);
-  }
-
-  const handleSpeedChange = (event) => {
-    const sliderValue = event.target.value
-    setSpeedState(sliderValue);
-
-    // considering the slider from 0 to 100 and the speed from 0 to 2000ms
-    const speedMs = getSpeed(sliderValue);
-
-    // Update interval
-    clearInterval(intervalRef.current);
-    intervalRef.current = setInterval( ()=> setNextGeneration(), speedMs);
-  }
-
+const Board = ({boardState, generationState, playIconState, speedState, onNextGeneration, onChangeSpeed, onPlay, onCellClick }) => {
 
   return (
     <>
@@ -66,16 +13,16 @@ const Board = () => {
           
           <div className={styles.controlsButtons}>           
 
-            <button className={styles.control} onClick={handlePlay} title='Play/Pause'>
+            <button className={styles.control} onClick={onPlay} title='Play/Pause'>
               <img src={playIconState} height="64px" />
             </button>
             
-            <button className={styles.control} onClick={handleNextGeneration} title='Next generation'>
+            <button className={styles.control} onClick={onNextGeneration} title='Next generation'>
               <img src={Step} height="40px"/>
             </button>
             
             <div className={styles.control} >
-              <input type="range" min="1" max="100" title='Speed' value={speedState} onChange={handleSpeedChange} className={styles.speed} />
+              <input type="range" min="1" max="100" title='Speed' value={speedState} onChange={onChangeSpeed} className={styles.speed} />
             </div>
             
           </div>
@@ -103,7 +50,7 @@ const Board = () => {
                     rowPosition={rowPosition}
                     columnPosition={columnPosition}
                     cellState={boardState[rowPosition][columnPosition]}
-                    onCellClick={handleCellClick}
+                    onCellClick={onCellClick}
                   />
                 </th>
               )}
@@ -119,6 +66,12 @@ const Board = () => {
 export default Board;
 
 Board.propTypes = {
-  rows: PropTypes.number,
-  columns: PropTypes.number,
+  boardState: PropTypes.array,
+  generationState: PropTypes.number, 
+  playIconState: PropTypes.object,
+  speedState: PropTypes.number,
+  onNextGeneration: PropTypes.func,
+  onChangeSpeed: PropTypes.func,
+  onPlay: PropTypes.func,
+  onCellClick: PropTypes.func,
 }
